+++ src/index.ts
@@ -7,10 +7,20 @@ export { default as UxAction } from "./components/Action";
 export { default as UxBubble } from "./components/Bubble";
 export { default as UxDot } from "./components/Dot";
 export { default as UxProgress } from "./components/Progress";
-export { default as UxSelect } from "./components/Select";
+export { default as UxSelect } from "./components/select/index";
+export { default as UxSelectOption } from "./components/select/SelectOption";
 export { default as UxDropdown } from "./components/Dropdown";
 export { default as UxToggle } from "./components/Toggle";
+export { default as UxTable } from "./components/table/Table";
+export { default as UxTableHeader } from "./components/table/TableHeader";
+export { default as UxTableBody } from "./components/table/TableBody";
+export { default as UxTableFooter } from "./components/table/TableFooter";
+export { default as UxTableHead } from "./components/table/TableHead";
+export { default as UxTableRow } from "./components/table/TableRow";
+export { default as UxTableDef } from "./components/table/TableDef";
+export { default as UxCheckbox } from "./components/Checkbox";
+export { default as UxInput } from "./components/Input";
+export { default as UxPagination } from "./components/Pagination";
 
 // Props
-export { SelectOptionProps } from "./types/select";
 export { DropdownOptionProps } from "./types/dropdown";