+++ src/types/checkbox/index.ts
@@ -1,0 +1,10 @@
+type CheckboxProps = {
+  variant?: "primary" | "secondary" | "dark" | "darker" | "darkest";
+  checked: boolean;
+  size?: "xs" | "sm" | "md" | "lg";
+  onAction: () => void;
+  rounded?: boolean;
+  cx?: string;
+};
+
+export { CheckboxProps };