+++ src/types/table/tdef.ts
@@ -1,0 +1,7 @@
+type TDefProps = {
+  children: JSX.Element | JSX.Element[] | string;
+  divideX?: boolean;
+  cx?: string;
+};
+
+export { TDefProps };