+++ src/types/table/tfooter.ts
@@ -1,0 +1,6 @@
+type TFooterProps = {
+  children: JSX.Element | JSX.Element[];
+  cx?: string;
+};
+
+export { TFooterProps };