+++ src/types/table/trow.ts
@@ -1,0 +1,3 @@
+type TRowProps = { children: JSX.Element | JSX.Element[]; cx?: string };
+
+export { TRowProps };